+++ src/components/plot/axes/plot/axis-plot-props-panel.tsx
@@ -4,7 +4,7 @@ import { DoubleNumericalInput } from '@/components/double-numerical-input'
 import { FontPopover } from '@/components/plot/font/font-popover'
 import { Input } from '@/components/shadcn/ui/themed/v2/input'
 import { produce } from 'immer'
-import { IPlotAddress, useAxes } from '../axes-store'
+import { IPlotAddress, useAxis } from '../axes-store'
 import { TickPlotPropsPopover } from './tick-plot-props-popover'
 
 export function AxisPlotPropsPanel({
@@ -14,13 +14,7 @@ export function AxisPlotPropsPanel({
   plotAddress: IPlotAddress
   title: string
 }) {
-  const { plots, updateAxis } = useAxes()
-
-  const { plotId, groupId, axisId } = plotAddress
-
-  const plot = plots[plotId]
-  const group = plot.groups[groupId]
-  const axis = group.axes[axisId]
+  const { axis, updateAxis } = useAxis(plotAddress)
 
   return (
     <>
@@ -29,10 +23,7 @@ export function AxisPlotPropsPanel({
         className="font-bold"
         checked={axis.style.show}
         onCheckedChange={(v) => {
-          updateAxis(
-            { plotId, groupId, axisId },
-            { style: { ...axis.style, show: v } }
-          )
+          updateAxis({ style: { ...axis.style, show: v } })
         }}
       />
 
@@ -40,26 +31,20 @@ export function AxisPlotPropsPanel({
         title="Title"
         checked={axis.style.title.show}
         onCheckedChange={(v) => {
-          updateAxis(
-            { plotId, groupId, axisId },
-            {
-              style: produce(axis.style, (draft) => {
-                draft.title.show = v
-              }),
-            }
-          )
+          updateAxis({
+            style: produce(axis.style, (draft) => {
+              draft.title.show = v
+            }),
+          })
         }}
       >
         <Input
           title="Title"
           value={axis.title}
           onTextChanged={(v) => {
-            updateAxis(
-              { plotId, groupId, axisId },
-              {
-                title: v,
-              }
-            )
+            updateAxis({
+              title: v,
+            })
           }}
           w="md"
         />
@@ -71,15 +56,12 @@ export function AxisPlotPropsPanel({
               textProps: axis.style.title,
               showEnabled: false,
               update: (f) =>
-                updateAxis(
-                  { plotId, groupId, axisId },
-                  {
-                    style: {
-                      ...axis.style,
-                      title: Object.assign({}, axis.style.title, f),
-                    },
-                  }
-                ),
+                updateAxis({
+                  style: {
+                    ...axis.style,
+                    title: Object.assign({}, axis.style.title, f),
+                  },
+                }),
             },
           ]}
         />
@@ -92,20 +74,14 @@ export function AxisPlotPropsPanel({
           limit={[-Infinity, Infinity]}
           dp={2}
           onNumChanged1={(v) => {
-            updateAxis(
-              { plotId, groupId, axisId },
-              {
-                domain: [v, axis.domain[1]],
-              }
-            )
+            updateAxis({
+              domain: [v, axis.domain[1]],
+            })
           }}
           onNumChanged2={(v) => {
-            updateAxis(
-              { plotId, groupId, axisId },
-              {
-                domain: [axis.domain[0], v],
-              }
-            )
+            updateAxis({
+              domain: [axis.domain[0], v],
+            })
           }}
         >
           -