+++ src/components/plot/axes/svg-axis-props.ts
@@ -15,8 +15,6 @@ import { deepmerge } from 'deepmerge-ts'
 
 export type TickLabel = string | number
 
-const MINOR_TICK_DIVISIONS = 5
-
 export interface ITickItem {
   v: number
   label?: string | undefined
@@ -116,8 +114,8 @@ export const DEFAULT_AXIS_CONFIG: IAxisConfig = {
   title: '',
 
   clip: true,
-  domain: [0, 1],
-  range: [0, 1],
+  domain: [0, 100],
+  range: [0, 100],
   style: {
     show: true,
     title: { ...DEFAULT_BOLD_TEXT_PROPS, offset: 30 },